refactor artemis and hermes boons
