--- db/boons/artemis.db.ts
@@ -1,52 +1,0 @@
-import { BoonData } from "models/types";
-
-export const DbBoonsArtemis: Record<string, BoonData> = {
-  pressurePoints: {
-    name: "Pressure Points",
-    key: "pressure-points",
-    god: "artemis",
-    element: "earth",
-  },
-
-  silverStreak: {
-    name: "Silver Streak",
-    key: "silver-streak",
-    god: "artemis",
-    element: "air",
-  },
-
-  firstBlood: {
-    name: "First Blood",
-    key: "first-blood",
-    god: "artemis",
-    element: "earth",
-  },
-
-  lethalSnare: {
-    name: "Lethal Snare",
-    key: "lethal-snare",
-    god: "artemis",
-    element: "earth",
-  },
-
-  easyShot: {
-    name: "Easy Shot",
-    key: "easy-shot",
-    god: "artemis",
-    element: "air",
-  },
-
-  supportFire: {
-    name: "Support Fire",
-    key: "support-fire",
-    god: "artemis",
-    element: "air",
-  },
-
-  deathWarrant: {
-    name: "Death Warrant",
-    key: "death-warrant",
-    god: "artemis",
-    element: "earth",
-  },
-};
