--- db/boons/hermes.db.ts
@@ -1,88 +1,0 @@
-import { BoonData } from "models/types";
-
-export const DbBoonsHermes: Record<string, BoonData> = {
-  tallOrder: {
-    name: "Tall Order",
-    key: "tall-order",
-    god: "hermes",
-    type: "infusion",
-  },
-
-  greaterEvasion: {
-    name: "Greater Evasion",
-    key: "greater-evasion",
-    god: "hermes",
-    element: "air",
-  },
-
-  savedBreath: {
-    name: "Saved Breath",
-    key: "saved-breath",
-    god: "hermes",
-    element: "earth",
-  },
-
-  nimbleMind: {
-    name: "Nimble Mind",
-    key: "nimble-mind",
-    god: "hermes",
-    element: "earth",
-  },
-
-  nimbleLimbs: {
-    name: "Nimble Limbs",
-    key: "nimble-limbs",
-    god: "hermes",
-    element: "earth",
-  },
-
-  midnightOil: {
-    name: "Midnight Oil",
-    key: "midnight-oil",
-    god: "hermes",
-    element: "air"
-  },
-
-  quickBuck: {
-    name: "Quick Buck",
-    key: "quick-buck",
-    god: "hermes",
-    element: "air",
-  },
-
-  hardTarget: {
-    name: "Hard Target",
-    key: "hard-target",
-    god: "hermes",
-    element: "air",
-  },
-
-  wittyRetort: {
-    name: "Witty Retort",
-    key: "witty-retort",
-    god: "hermes",
-    element: "earth",
-  },
-
-  nitroBoost: {
-    name: "Nitro Boost",
-    key: "nitro-boost",
-    god: "hermes",
-    element: "fire",
-  },
-
-  meanStreak: {
-    name: "Mean Streak",
-    key: "mean-streak",
-    god: "hermes",
-    element: "air",
-  },
-
-  closeCall: {
-    name: "Close Call",
-    key: "close-call",
-    god: "hermes",
-    element: "air",
-    type: "legendary",
-  },
-};
